+++ NEWS.md
@@ -17,6 +17,7 @@
 * Improved and simplified scheme plotting in `plot_scheme()` function. Removed {ggstance} from package dependencies
 * Added `label_size` and `font_size` arguments to `plot_scheme()` and `plot_progression_table()` functions
 * Removed default progression table from `generate_progression_table()`, `create_example()`, `plot_progression_table()` functions
+* Added `plot_vertical()` function for plotting vertical plan
 
 # STMr 0.1.3